requirements.txt
aws-cdk.core==1.121.0

aws-cdk.aws-lambda==1.121.0
aws_cdk.aws_lambda_python==1.121.0
aws-cdk.aws_apigateway==1.121.0

